change assertions

+++ tests/tests_pytorch/plugins/precision/test_fsdp.py
@@ -83,14 +83,14 @@ def test_fsdp_precision_scaler_with_bf16():
 @RunIf(min_torch="1.12", min_cuda_gpus=1)
 def test_fsdp_precision_forward_context():
     """Test to ensure that the context manager correctly is set to bfloat16."""
-    precision = FSDPPrecisionPlugin(precision="16-mixed")
+    precision = FSDPPrecisionPlugin(precision="16-true")
     assert isinstance(precision.scaler, torch.cuda.amp.GradScaler)
     assert torch.get_default_dtype() == torch.float32
     with precision.forward_context():
         # check with str due to a bug upstream: https://github.com/pytorch/pytorch/issues/65786
         assert str(torch.get_autocast_gpu_dtype()) in ("torch.float16", "torch.half")
 
-    precision = FSDPPrecisionPlugin(precision="bf16-mixed")
+    precision = FSDPPrecisionPlugin(precision="bf16-true")
     assert precision.scaler is None
     with precision.forward_context():
         # check with str due to a bug upstream: https://github.com/pytorch/pytorch/issues/65786
